6.20

+++ src/test/java/com/lulu/usercenter/algorithmTest.java
@@ -12,8 +12,8 @@ public class algorithmTest {
         List<String> list = Arrays.asList("java", "大一", "男");
         List<String> list1 = Arrays.asList("java", "大二", "男");
         List<String> list2 = Arrays.asList("c++", "大二", "女");
-        int i1 = AlgorithmUtils.minDistance(list, list1);
-        int i2 = AlgorithmUtils.minDistance(list, list2);
+        Long i1 = AlgorithmUtils.minDistance(list, list1);
+        Long i2 = AlgorithmUtils.minDistance(list, list2);
         System.out.println(i1);
         System.out.println(i2);
     }
